structural: Share worker trust repository fixtures
